requirements/test.txt
tox
pytest==2.7.3
pytest-django==2.8.0
factory_boy==2.5.2
